pyproject.toml
[tool.poetry]
name = "knonwledge-discovery"
version = "0.1.0"
description = ""
authors = ["Cephas Kingo'ri <you@example.com>", "Jerim Kaura <you@example.com>", "Stephen Mwangi <you@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
python-javabridge = "^4.0.3"
python-weka-wrapper3 = "^0.2.7"
matplotlib = "^3.5.1"
pandas = "^1.4.1"
scikit-learn = "^1.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
